backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
ta>=0.11.0
anthropic>=0.42.0
feedparser>=6.0.0
pydantic>=2.7.0
python-dotenv>=1.0.0
vectorbt>=0.26.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
gitpython>=3.1.0
pytest>=8.0.0
httpx>=0.27.0
fredapi>=0.5.0
